requirements.txt
Flask==2.3.2
Requests==2.31.0
Bootstrap_Flask==2.2.0
gunicorn==22.0.0
psycopg2-binary==2.9.9
Pillow==10.3.0
readwise==1.1.0
flask-caching==2.3.0
python-dotenv==0.19.1
black==24.10.0